move helper functions to metrics pkg
