+++ test/performance/performance.go
@@ -17,22 +17,16 @@ limitations under the License.
 package performance
 
 import (
-	"context"
 	"fmt"
-	"log"
 	"os"
 	"path"
 	"strings"
 	"testing"
 	"time"
 
-	"k8s.io/apimachinery/pkg/labels"
-	deploymentinformer "knative.dev/pkg/injection/informers/kubeinformers/appsv1/deployment"
 	pkgTest "knative.dev/pkg/test"
 	"knative.dev/pkg/test/logging"
 	"knative.dev/pkg/test/zipkin"
-	netv1alpha1 "knative.dev/serving/pkg/apis/networking/v1alpha1"
-	sksinformer "knative.dev/serving/pkg/client/injection/informers/networking/v1alpha1/serverlessservice"
 	"knative.dev/serving/test"
 	"knative.dev/test-infra/shared/common"
 	"knative.dev/test-infra/shared/prometheus"
@@ -157,92 +151,4 @@ func sanitizedURL(endpoint string) string {
 		return httpPrefix + endpoint
 	}
 	return endpoint
-}
-
-// DeploymentStatus is a struct that wraps the status of a deployment.
-type DeploymentStatus struct {
-	DesiredReplicas int32
-	ReadyReplicas   int32
-	// Time is the time when the status is fetched
-	Time time.Time
-}
-
-// FetchDeploymentStatus creates a channel that can return the up-to-date DeploymentStatus periodically.
-func FetchDeploymentStatus(
-	ctx context.Context, namespace string, selector labels.Selector,
-	duration time.Duration, stop <-chan struct{},
-) <-chan DeploymentStatus {
-	dl := deploymentinformer.Get(ctx).Lister()
-	ch := make(chan DeploymentStatus)
-	startTick(duration, stop, func(t time.Time) error {
-		// Overlay the desired and ready pod counts.
-		deployments, err := dl.Deployments(namespace).List(selector)
-		if err != nil {
-			log.Printf("Error listing deployments: %v", err)
-			return err
-		}
-
-		for _, d := range deployments {
-			ds := DeploymentStatus{
-				DesiredReplicas: *d.Spec.Replicas,
-				ReadyReplicas:   d.Status.ReadyReplicas,
-				Time:            t,
-			}
-			ch <- ds
-		}
-		return nil
-	})
-
-	return ch
-}
-
-// ServerlessServiceStatus is a struct that wraps the status of a serverless service.
-type ServerlessServiceStatus struct {
-	Mode netv1alpha1.ServerlessServiceOperationMode
-	// Time is the time when the status is fetched
-	Time time.Time
-}
-
-// FetchSKSMode creates a channel that can return the up-to-date ServerlessServiceOperationMode periodically.
-func FetchSKSMode(
-	ctx context.Context, namespace string, selector labels.Selector,
-	duration time.Duration, stop <-chan struct{},
-) <-chan ServerlessServiceStatus {
-	sksl := sksinformer.Get(ctx).Lister()
-	ch := make(chan ServerlessServiceStatus)
-	startTick(duration, stop, func(t time.Time) error {
-		// Overlay the SKS "mode".
-		skses, err := sksl.ServerlessServices(namespace).List(selector)
-		if err != nil {
-			log.Printf("Error listing serverless services: %v", err)
-			return err
-		}
-		for _, sks := range skses {
-			skss := ServerlessServiceStatus{
-				Mode: sks.Spec.Mode,
-				Time: t,
-			}
-			ch <- skss
-		}
-		return nil
-	})
-
-	return ch
-}
-
-func startTick(duration time.Duration, stop <-chan struct{}, action func(t time.Time) error) {
-	ticker := time.NewTicker(duration)
-	go func() {
-		for {
-			select {
-			case t := <-ticker.C:
-				if err := action(t); err != nil {
-					break
-				}
-			case <-stop:
-				ticker.Stop()
-				return
-			}
-		}
-	}()
 }
